MODWRKFLOW-2: Restructure enumerations, relocating them into another folder.
Separate the enumerations from the models with the goal of helping make managing the different models somewhat easier.
