+++ components/src/main/java/org/folio/rest/workflow/enums/CompressFileContainer.java
@@ -1,4 +1,4 @@
-package org.folio.rest.workflow.model;
+package org.folio.rest.workflow.enums;
 
 public enum CompressFileContainer {
   NONE, TAR
