+++ components/src/main/java/org/folio/rest/workflow/enums/CompressFileFormat.java
@@ -1,4 +1,4 @@
-package org.folio.rest.workflow.model;
+package org.folio.rest.workflow.enums;
 
 public enum CompressFileFormat {
   BZIP2, GZIP, ZIP
