+++ components/src/main/java/org/folio/rest/workflow/enums/Direction.java
@@ -1,4 +1,4 @@
-package org.folio.rest.workflow.model;
+package org.folio.rest.workflow.enums;
 
 public enum Direction {
   UNSPECIFIED("Unspecified"),
