+++ components/src/main/java/org/folio/rest/workflow/enums/DirectoryAction.java
@@ -1,4 +1,4 @@
-package org.folio.rest.workflow.model;
+package org.folio.rest.workflow.enums;
 
 public enum DirectoryAction {
   READ_NEXT, DELETE_NEXT, LIST, WRITE
