+++ components/src/main/java/org/folio/rest/workflow/enums/ScriptType.java
@@ -1,4 +1,4 @@
-package org.folio.rest.workflow.model;
+package org.folio.rest.workflow.enums;
 
 public enum ScriptType {
   GROOVY("groovy"),
