+++ components/src/main/java/org/folio/rest/workflow/enums/SftpOp.java
@@ -1,0 +1,5 @@
+package org.folio.rest.workflow.enums;
+
+public enum SftpOp {
+  GET, PUT
+}
