+++ components/src/main/java/org/folio/rest/workflow/enums/StartEventType.java
@@ -1,4 +1,4 @@
-package org.folio.rest.workflow.model;
+package org.folio.rest.workflow.enums;
 
 public enum StartEventType {
   MESSAGE_CORRELATION, SCHEDULED, SIGNAL, NONE
