+++ components/src/main/java/org/folio/rest/workflow/model/EmbeddedProcessor.java
@@ -6,6 +6,7 @@ import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
+import org.folio.rest.workflow.enums.ScriptType;
 
 @Embeddable
 public class EmbeddedProcessor {
