+++ components/src/main/java/org/folio/rest/workflow/model/EmbeddedVariable.java
@@ -8,6 +8,7 @@ import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
+import org.folio.rest.workflow.enums.VariableType;
 
 @Embeddable
 public class EmbeddedVariable {
