--- components/src/main/java/org/folio/rest/workflow/model/SftpOp.java
@@ -1,5 +1,0 @@
-package org.folio.rest.workflow.model;
-
-public enum SftpOp {
-  GET, PUT
-}
